requirements-dev.txt
bandit
black
flake8
isort
mypy
pyyaml
ruff
yamllint
types-PyYAML
